Refactor Inventory and Radionuclide classes into separate .py files
